+++ PickupPointBackEnd/src/main/java/com/tqs/pickuppointbackend/controller/PickupPointController.java
@@ -1,6 +1,9 @@
 package com.tqs.pickuppointbackend.controller;
 
 
+import com.tqs.pickuppointbackend.constants.UserType;
+import com.tqs.pickuppointbackend.controller.model.UpdateRequest;
+import com.tqs.pickuppointbackend.service.AuthenticationService;
 import lombok.extern.log4j.Log4j2;
 
 import java.util.List;
@@ -9,13 +12,7 @@ import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import com.tqs.pickuppointbackend.exceptions.ResourceNotFoundException;
 import com.tqs.pickuppointbackend.model.PickupPoint;
@@ -29,59 +26,86 @@ public class PickupPointController {
     @Autowired
     PickupPointService pickupPointService;
 
-    @GetMapping("/pickuppoints")
-    public ResponseEntity<List<PickupPoint>> getPickupPoints(){
+    @Autowired
+    private AuthenticationService authenticationService;
 
-        return ResponseEntity.ok().body(pickupPointService.getPickupPoints());
+    @GetMapping("/pickuppoints")
+    public ResponseEntity<List<PickupPoint>> getPickupPoints(@RequestHeader("Authorization") String token) throws ResourceNotFoundException {
+
+        if (authenticationService.hasAcess(token, UserType.CLIENT)){
+            return ResponseEntity.ok().body(pickupPointService.getPickupPoints());
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
 
     }
 
     @GetMapping("/pickuppoint/{id}")
-    public ResponseEntity<PickupPoint> getPickupScheduleById(@PathVariable(value = "id") long id) throws ResourceNotFoundException {
+    public ResponseEntity<PickupPoint> getPickupScheduleById(@PathVariable(value = "id") long id, @RequestHeader("Authorization") String token) throws ResourceNotFoundException {
 
-        return ResponseEntity.ok().body(pickupPointService.getPickupPointById(id));
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok().body(pickupPointService.getPickupPointById(id));
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
 
     }
 
     @GetMapping("/pickuppoint/user/{user_id}")
-    public ResponseEntity<PickupPoint> getPickupPointByuserId(@PathVariable(value = "user_id") long id) throws ResourceNotFoundException {
+    public ResponseEntity<PickupPoint> getPickupPointByuserId(@PathVariable(value = "user_id") long id, @RequestHeader("Authorization") String token) throws ResourceNotFoundException {
 
-        return ResponseEntity.ok().body(pickupPointService.getPickupPointByUserId(id));
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok().body(pickupPointService.getPickupPointByUserId(id));
+        }
+        throw new ResourceNotFoundException("The user don't have acess");
 
     }
 
     @GetMapping("/pickuppoints/available")
-    public ResponseEntity<List<PickupPoint>> getPickupPointById() throws ResourceNotFoundException {
+    public ResponseEntity<List<PickupPoint>> getPickupPointById(@RequestHeader("Authorization") String token) throws ResourceNotFoundException {
 
-        return ResponseEntity.ok().body(pickupPointService.getAvailablePickupPoints());
-
+        if (authenticationService.hasAcess(token, UserType.ADMIN)){
+            return ResponseEntity.ok().body(pickupPointService.getAvailablePickupPoints());
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
     }
 
     @GetMapping("/pickuppoints/nonavailable")
-    public ResponseEntity<List<PickupPoint>> getPickupPointNonAvailable() throws ResourceNotFoundException {
-
-        return ResponseEntity.ok().body(pickupPointService.getNonAvailablePickupPoints());
+    public ResponseEntity<List<PickupPoint>> getPickupPointNonAvailable(@RequestHeader("Authorization") String token) throws ResourceNotFoundException {
+        if (authenticationService.hasAcess(token, UserType.ADMIN)){
+            return ResponseEntity.ok().body(pickupPointService.getNonAvailablePickupPoints());
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
 
     }
 
     @PostMapping("/pickuppoint")
-    public ResponseEntity<PickupPoint> addPickupPoint(@Valid @RequestBody PickupPointDTO pickupPointDTO) throws ResourceNotFoundException {
+    public ResponseEntity<PickupPoint> addPickupPoint(@Valid @RequestBody PickupPointDTO pickupPointDTO,  @RequestHeader("Authorization") String token) throws ResourceNotFoundException {
 
-        return ResponseEntity.ok().body(pickupPointService.addPickupPoint(pickupPointDTO));
-
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok().body(pickupPointService.addPickupPoint(pickupPointDTO));
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
     }
 
     @PutMapping("/pickuppoint/{id}")
-    public ResponseEntity<PickupPoint> updatePickupPoint(@PathVariable(value = "id") Long id, @Valid @RequestBody PickupPointDTO pickupPointDTO) throws ResourceNotFoundException {
-        System.out.println(id);
-        return ResponseEntity.ok(pickupPointService.updatePickupPoint(pickupPointDTO));
+    public ResponseEntity<PickupPoint> updatePickupPoint(@PathVariable(value = "id") Long id, @Valid @RequestBody UpdateRequest request, @RequestHeader("Authorization") String token) throws ResourceNotFoundException {
+        log.info(id);
+        log.info(request);
+
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok(pickupPointService.updatePickupPoint(id, request));
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
+
     }
 
     @DeleteMapping("/pickuppoint/{id}")
-    public ResponseEntity<PickupPoint> deletePickupPointById(@PathVariable(value = "id") Long id) throws ResourceNotFoundException {
+    public ResponseEntity<PickupPoint> deletePickupPointById(@PathVariable(value = "id") Long id,  @RequestHeader("Authorization") String token) throws ResourceNotFoundException {
 
-        return ResponseEntity.ok().body(pickupPointService.deletePickupPointById(id));
-    
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok().body(pickupPointService.deletePickupPointById(id));
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
+
     }
     
 }
